Extract the nauro command resolver into cli/nauro_command.py
Move the MCP/hook command resolution and validation seam out of
cli/utils.py and cli/commands/setup.py into a dedicated
cli/nauro_command.py. The probe seam (probe_nauro_command), the
durability heuristic (_is_durable_install_path), the interpreter-sibling
lookup, the two resolver-warning constants, the cached
_find_nauro_command / _resolve_nauro_command entrypoints, and the Codex
hook resolver now live together in one module, so callers no longer
reach across utils and setup to resolve the recorded command.

This is a pure relocation with no behavior change. Every moved body is
byte-identical to its origin, except that _resolve_nauro_command and
_find_nauro_codex_hook_command drop the now-redundant cli_utils. prefix
on probe_nauro_command and _is_durable_install_path, which are colocated
in the new module. status and adopt retarget to nauro_command; the
autouse test fixture and the resolver tests patch nauro_command in place
of utils and setup.
