+++ packages/nauro/src/nauro/cli/commands/setup.py
@@ -10,10 +10,7 @@ Subcommands:
 
 from __future__ import annotations
 
-import functools
 import json
-import shutil
-import sys
 from dataclasses import dataclass
 from pathlib import Path
 
@@ -22,9 +19,7 @@ import typer
 from tomlkit.exceptions import ParseError as TOMLParseError
 from tomlkit.items import InlineTable
 
-from nauro.cli import utils as cli_utils
 from nauro.cli._codex_hooks import (
-    _CODEX_HOOK_PROBE_ARGS,
     _CodexHookConfigError,
     _format_codex_hooks,
     _parse_codex_hooks,
@@ -41,6 +36,7 @@ from nauro.cli.integrations.skills import (
     materialize_skills_cursor_for_repo,
 )
 from nauro.cli.integrations.user_scope import _registered_project_keys, _user_scope_safe_to_clear
+from nauro.cli.nauro_command import _find_nauro_codex_hook_command, _find_nauro_command
 from nauro.cli.utils import _resolve_project_entry, resolve_target_project
 from nauro.store._atomic import atomic_write_text
 from nauro.store.registry import get_repo_paths
@@ -56,100 +52,6 @@ setup_app = typer.Typer(help="Configure tool integrations.")
 # against the local store, so users don't have to wait for an agent
 # restart to see Nauro do something useful.
 CHECK_HINT_LINE = 'Try it now from this shell: nauro check-decision "<approach>"'
-
-
-def _interpreter_sibling_candidate() -> str | None:
-    """Return the absolute path to a ``nauro`` console script next to the running
-    interpreter, or None when there isn't one.
-
-    This is the install the user actually invoked, which pipx/uv-tool layouts
-    keep off the PATH that GUI-launched agents see — recording its absolute path
-    is what makes the spawned stdio server and per-turn hook independent of the
-    agent's launch environment.
-    """
-    bindir = Path(sys.executable).parent
-    for name in ("nauro", "nauro.exe"):
-        candidate = bindir / name
-        if candidate.is_file():
-            return str(candidate)
-    return None
-
-
-_FRAGILE_COMMAND_WARNING = (
-    "WARNING: recording nauro from a project virtualenv ({command}).\n"
-    "  This path breaks if the repo's virtualenv is rebuilt, moved, or "
-    "corrupted, silently killing Nauro's MCP server and hooks. Install nauro "
-    "durably (pipx install nauro, or uv tool install nauro) and re-run "
-    "'nauro setup all'."
-)
-
-_UNRESOLVED_COMMAND_WARNING = (
-    "WARNING: could not validate a working nauro; recorded '{command}'.\n"
-    "  Nauro's MCP server and hooks will not work until nauro is installed on a "
-    "durable PATH (pipx install nauro, or uv tool install nauro), then re-run "
-    "'nauro setup all'."
-)
-
-
-@functools.cache
-def _find_nauro_command() -> str:
-    """Resolve — and cache for the process — the nauro entrypoint recorded into
-    MCP and hook configs.
-
-    Cached so `setup all` validates the entrypoint once rather than once per
-    sink (five subprocess probes collapse to one). Warnings surface on the
-    cache-miss resolution only; tests reset via
-    ``_find_nauro_command.cache_clear()``.
-    """
-    return _resolve_nauro_command()
-
-
-def _resolve_nauro_command() -> str:
-    """Pick the nauro entrypoint to record into MCP/hook configs.
-
-    Prefers a validated, durable install so the recorded command keeps working
-    after a project virtualenv is rebuilt, moved, or corrupted (the observed
-    failure: a ``uv run`` / ``.venv``-invoked setup recorded a fragile
-    repo-venv path that later died). Resolution order:
-
-      1. Interpreter-sibling that both runs and looks durable — the fast path,
-         byte-identical to the historical behavior for pipx/uv-tool/desktop.
-      2. Otherwise a PATH-resolved absolute shim that runs and looks durable —
-         diverts away from a dead or fragile project venv.
-      3. Otherwise the sibling if it merely runs (fragile but working) —
-         recorded with a loud warning naming the project-venv fragility.
-      4. Otherwise the best absolute path we have (else bare ``nauro``), with a
-         loud warning that MCP will not work until nauro is on a durable PATH.
-
-    An absolute path is always preferred over bare ``nauro``; bare ``nauro`` is
-    only the terminal fallback, because GUI-launched agents start with an empty
-    PATH. Durability checks run before the (subprocess) probe so a non-durable
-    candidate short-circuits without spawning.
-    """
-    sibling = _interpreter_sibling_candidate()
-    which = shutil.which("nauro")
-
-    if (
-        sibling is not None
-        and cli_utils._is_durable_install_path(sibling)
-        and cli_utils.probe_nauro_command(sibling)
-    ):
-        return sibling
-
-    if (
-        which is not None
-        and cli_utils._is_durable_install_path(which)
-        and cli_utils.probe_nauro_command(which)
-    ):
-        return which
-
-    if sibling is not None and cli_utils.probe_nauro_command(sibling):
-        typer.echo(_FRAGILE_COMMAND_WARNING.format(command=sibling), err=True)
-        return sibling
-
-    fallback = sibling or which or "nauro"
-    typer.echo(_UNRESOLVED_COMMAND_WARNING.format(command=fallback), err=True)
-    return fallback
 
 
 def _configure_json_mcp(
@@ -746,35 +648,6 @@ def _remove_hook_entry(settings_path: Path, settings: dict, repo: Path) -> str:
 
 def _codex_hooks_path(repo: Path) -> Path:
     return repo / ".codex" / "hooks.json"
-
-
-@functools.cache
-def _find_nauro_codex_hook_command() -> str | None:
-    command = _find_nauro_command()
-    if cli_utils.probe_nauro_command(command, args=_CODEX_HOOK_PROBE_ARGS):
-        return command
-
-    sibling = _interpreter_sibling_candidate()
-    if (
-        sibling is not None
-        and sibling != command
-        and cli_utils.probe_nauro_command(sibling, args=_CODEX_HOOK_PROBE_ARGS)
-    ):
-        typer.echo(
-            f"WARNING: '{command}' does not support Codex bootstrap hooks. "
-            f"Recording the current Nauro install at '{sibling}' instead. "
-            "Update the durable Nauro install and re-run 'nauro setup all --with-hooks'.",
-            err=True,
-        )
-        return sibling
-
-    typer.echo(
-        "WARNING: no installed Nauro command supports Codex bootstrap hooks. "
-        "Codex hook wiring was skipped; update Nauro and re-run "
-        "'nauro setup all --with-hooks'.",
-        err=True,
-    )
-    return None
 
 
 def materialize_hooks_codex(repo: Path, *, remove: bool) -> str:
