+++ packages/nauro/tests/test_command_resolution.py
@@ -21,12 +21,11 @@ import subprocess
 
 import pytest
 
-from nauro.cli import utils as cli_utils
-from nauro.cli.commands import setup as setup_mod
+from nauro.cli import nauro_command
 
 # Real implementations captured before the autouse fixture patches them.
-_REAL_PROBE = cli_utils.probe_nauro_command
-_REAL_DURABLE = cli_utils._is_durable_install_path
+_REAL_PROBE = nauro_command.probe_nauro_command
+_REAL_DURABLE = nauro_command._is_durable_install_path
 
 
 # ── probe_nauro_command ────────────────────────────────────────────────────────
@@ -38,12 +37,12 @@ class _FakeProc:
 
 
 def test_probe_true_on_exit_zero(monkeypatch):
-    monkeypatch.setattr(cli_utils.subprocess, "run", lambda *a, **k: _FakeProc(0))
+    monkeypatch.setattr(nauro_command.subprocess, "run", lambda *a, **k: _FakeProc(0))
     assert _REAL_PROBE("nauro") is True
 
 
 def test_probe_false_on_nonzero_exit(monkeypatch):
-    monkeypatch.setattr(cli_utils.subprocess, "run", lambda *a, **k: _FakeProc(1))
+    monkeypatch.setattr(nauro_command.subprocess, "run", lambda *a, **k: _FakeProc(1))
     assert _REAL_PROBE("nauro") is False
 
 
@@ -51,7 +50,7 @@ def test_probe_false_on_missing_binary(monkeypatch):
     def boom(*a, **k):
         raise FileNotFoundError("nauro")
 
-    monkeypatch.setattr(cli_utils.subprocess, "run", boom)
+    monkeypatch.setattr(nauro_command.subprocess, "run", boom)
     assert _REAL_PROBE("/gone/nauro") is False
 
 
@@ -59,7 +58,7 @@ def test_probe_false_on_timeout(monkeypatch):
     def boom(*a, **k):
         raise subprocess.TimeoutExpired(cmd="nauro", timeout=1.5)
 
-    monkeypatch.setattr(cli_utils.subprocess, "run", boom)
+    monkeypatch.setattr(nauro_command.subprocess, "run", boom)
     assert _REAL_PROBE("nauro") is False
 
 
@@ -70,7 +69,7 @@ def test_probe_invokes_version_subcommand(monkeypatch):
         seen["cmd"] = cmd
         return _FakeProc(0)
 
-    monkeypatch.setattr(cli_utils.subprocess, "run", capture)
+    monkeypatch.setattr(nauro_command.subprocess, "run", capture)
     _REAL_PROBE("/opt/nauro")
     assert seen["cmd"] == ["/opt/nauro", "--version"]
 
@@ -82,7 +81,7 @@ def test_probe_accepts_a_specific_subcommand(monkeypatch):
         seen["cmd"] = cmd
         return _FakeProc(0)
 
-    monkeypatch.setattr(cli_utils.subprocess, "run", capture)
+    monkeypatch.setattr(nauro_command.subprocess, "run", capture)
     result = _REAL_PROBE("/opt/nauro", args=("hook", "codex-bootstrap", "--help"))
 
     assert result is True
@@ -121,10 +120,10 @@ def test_is_durable_install_path(path, expected):
 
 def _wire_resolver(monkeypatch, *, sibling, which, probe, durable):
     """Point the resolver at controlled candidates and seam functions."""
-    monkeypatch.setattr(setup_mod, "_interpreter_sibling_candidate", lambda: sibling)
-    monkeypatch.setattr(setup_mod.shutil, "which", lambda name: which)
-    monkeypatch.setattr(cli_utils, "probe_nauro_command", probe)
-    monkeypatch.setattr(cli_utils, "_is_durable_install_path", durable)
+    monkeypatch.setattr(nauro_command, "_interpreter_sibling_candidate", lambda: sibling)
+    monkeypatch.setattr(nauro_command.shutil, "which", lambda name: which)
+    monkeypatch.setattr(nauro_command, "probe_nauro_command", probe)
+    monkeypatch.setattr(nauro_command, "_is_durable_install_path", durable)
 
 
 def test_resolver_records_durable_sibling_without_warning(monkeypatch, capsys):
@@ -140,7 +139,7 @@ def test_resolver_records_durable_sibling_without_warning(monkeypatch, capsys):
         probe=lambda cmd, **k: True,
         durable=lambda p: True,
     )
-    assert setup_mod._resolve_nauro_command() == sibling
+    assert nauro_command._resolve_nauro_command() == sibling
     assert capsys.readouterr().err == ""
 
 
@@ -155,7 +154,7 @@ def test_resolver_diverts_to_which_when_sibling_dead(monkeypatch, capsys):
         probe=lambda cmd, **k: cmd == which,  # sibling crashes on import
         durable=lambda p: p == which,  # sibling is fragile, which is durable
     )
-    assert setup_mod._resolve_nauro_command() == which
+    assert nauro_command._resolve_nauro_command() == which
     # Silent divert — no warning when a durable replacement is found.
     assert capsys.readouterr().err == ""
 
@@ -170,7 +169,7 @@ def test_resolver_records_fragile_sibling_with_warning(monkeypatch, capsys):
         probe=lambda cmd, **k: True,  # sibling runs
         durable=lambda p: False,  # ...but is fragile
     )
-    assert setup_mod._resolve_nauro_command() == sibling
+    assert nauro_command._resolve_nauro_command() == sibling
     err = capsys.readouterr().err
     assert "WARNING" in err
     assert "virtualenv" in err
@@ -187,7 +186,7 @@ def test_resolver_falls_back_with_loud_warning_when_nothing_runs(monkeypatch, ca
         probe=lambda cmd, **k: False,  # nothing runs
         durable=lambda p: False,
     )
-    assert setup_mod._resolve_nauro_command() == sibling
+    assert nauro_command._resolve_nauro_command() == sibling
     err = capsys.readouterr().err
     assert "WARNING" in err
     assert "will not work" in err
@@ -202,7 +201,7 @@ def test_resolver_bare_nauro_only_as_last_resort(monkeypatch, capsys):
         probe=lambda cmd, **k: False,
         durable=lambda p: False,
     )
-    assert setup_mod._resolve_nauro_command() == "nauro"
+    assert nauro_command._resolve_nauro_command() == "nauro"
     assert "WARNING" in capsys.readouterr().err
 
 
@@ -216,7 +215,7 @@ def test_resolver_never_prefers_bare_over_absolute(monkeypatch, capsys):
         probe=lambda cmd, **k: False,  # neither validates
         durable=lambda p: False,
     )
-    result = setup_mod._resolve_nauro_command()
+    result = nauro_command._resolve_nauro_command()
     assert result != "nauro"
     assert result == sibling  # best absolute path retained
     capsys.readouterr()
@@ -234,16 +233,16 @@ def test_find_nauro_command_memoizes_resolution(monkeypatch):
         return True
 
     monkeypatch.setattr(
-        setup_mod, "_interpreter_sibling_candidate", lambda: "/opt/pipx/venvs/nauro/bin/nauro"
+        nauro_command, "_interpreter_sibling_candidate", lambda: "/opt/pipx/venvs/nauro/bin/nauro"
     )
-    monkeypatch.setattr(setup_mod.shutil, "which", lambda name: None)
-    monkeypatch.setattr(cli_utils, "probe_nauro_command", counting_probe)
-    monkeypatch.setattr(cli_utils, "_is_durable_install_path", lambda p: True)
-    setup_mod._find_nauro_command.cache_clear()
+    monkeypatch.setattr(nauro_command.shutil, "which", lambda name: None)
+    monkeypatch.setattr(nauro_command, "probe_nauro_command", counting_probe)
+    monkeypatch.setattr(nauro_command, "_is_durable_install_path", lambda p: True)
+    nauro_command._find_nauro_command.cache_clear()
 
-    first = setup_mod._find_nauro_command()
+    first = nauro_command._find_nauro_command()
     after_first = len(calls)
-    second = setup_mod._find_nauro_command()
+    second = nauro_command._find_nauro_command()
 
     assert first == second == "/opt/pipx/venvs/nauro/bin/nauro"
     assert len(calls) == after_first  # no re-probe on the cached call
@@ -261,14 +260,14 @@ def test_setup_all_resolves_command_once(tmp_path, monkeypatch):
 
     calls: list[str] = []
     monkeypatch.setattr(
-        setup_mod, "_interpreter_sibling_candidate", lambda: "/opt/pipx/venvs/nauro/bin/nauro"
+        nauro_command, "_interpreter_sibling_candidate", lambda: "/opt/pipx/venvs/nauro/bin/nauro"
     )
-    monkeypatch.setattr(setup_mod.shutil, "which", lambda name: None)
+    monkeypatch.setattr(nauro_command.shutil, "which", lambda name: None)
     monkeypatch.setattr(
-        cli_utils, "probe_nauro_command", lambda cmd, **k: calls.append(cmd) or True
+        nauro_command, "probe_nauro_command", lambda cmd, **k: calls.append(cmd) or True
     )
-    monkeypatch.setattr(cli_utils, "_is_durable_install_path", lambda p: True)
-    setup_mod._find_nauro_command.cache_clear()
+    monkeypatch.setattr(nauro_command, "_is_durable_install_path", lambda p: True)
+    nauro_command._find_nauro_command.cache_clear()
 
     setup_all_surfaces([repo1, repo2], remove=False)
